requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
meraki==1.20.0
multidict==6.0.2
paho-mqtt==1.6.1
Pillow==9.1.0
psutil==5.9.0
requests==2.27.1
urllib3==1.26.9
yarl==1.7.2
